+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   - [x] Enabled in `production` environment (displayed in log file)
 - Allow static files to be served from the `static` directory
 - Change file name format for log files (to `YYYY-MM-DD_hh-mm-ss`)
+- Add file support for static XML files
 
 #### Minor changes